requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
cryptography>=42.0,<43.0
Pillow>=10.0,<11.0
django-ratelimit>=4.1,<5.0
django-filter>=24.0,<25.0
